backend/pyproject.toml
[project]
name = "clearmoney-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "Django==6.0.3",
    "psycopg[binary]==3.2.6",
    "gunicorn==23.0.0",
    "django-htmx==1.27.0",
    "dj-database-url==3.1.0",
    "django-ratelimit==4.1.0",  # Rate limiting decorators — like Go's token bucket middleware
    "whitenoise>=6.12.0",
    "python-json-logger>=4.0.0",
    "python-dateutil>=2.9.0",
    "resend>=2.0.0",
    "weasyprint>=68.1",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-django==4.9.0",
    "pytest-mock==3.14.0",
    "pytest-xdist==3.6.1",
    "pytest-cov==6.0.0",
    "factory-boy==3.3.1",
    "freezegun>=1.5.1",
    "ruff==0.15.6",
    "mypy==1.19.1",
    "django-stubs[compatible-mypy]==6.0.1",
    "types-python-dateutil>=2.9.0",
    "django-ai-boost>=0.8.0",
    "import-linter>=2.11",
    "basedpyright>=1.0.0",
    "lsp>=0.1.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "clearmoney.settings"
python_files = ["tests.py", "test_*.py", "*_test.py"]
addopts = "--reuse-db -p no:warnings"

# ---------- Coverage ----------
[tool.coverage.run]
source = [
    "clearmoney",
    "core",
    "auth_app",
    "dashboard",
    "accounts",
    "transactions",
    "people",
    "budgets",
    "virtual_accounts",
    "recurring",
    "investments",
    "installments",
    "exchange_rates",
    "categories",
    "push",
    "jobs",
    "settings_app",
    "reports",
]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/test_*.py",
    "conftest.py",
    "manage.py",
    "clearmoney/wsgi.py",
    "clearmoney/asgi.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "pass$",
]
fail_under = 60

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]  # line length handled by formatter

[tool.ruff.lint.isort]
known-first-party = ["clearmoney", "core", "settings_app", "reports", "dashboard", "accounts"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
# Require annotations on all functions
disallow_untyped_defs = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
# Packages without stubs (factory-boy, psycopg cursors, etc.)
ignore_missing_imports = true

# Relax rules for test files — pytest fixtures use dynamic types
[[tool.mypy.overrides]]
module = ["conftest", "*.tests", "*.tests.*"]
disallow_untyped_defs = false

[tool.django-stubs]
django_settings_module = "clearmoney.settings"

# ── Import Linter (enforce modular boundaries) ────────────────────────────────

[tool.importlinter]
root_packages = [
    "accounts", "auth_app", "budgets", "categories", "core",
    "dashboard", "exchange_rates", "investments", "jobs",
    "people", "push", "recurring", "reports",
    "settings_app", "transactions", "virtual_accounts",
]

[[tool.importlinter.contracts]]
name = "Leaf modules must not import from aggregators"
type = "forbidden"
source_modules = [
    "accounts", "auth_app", "budgets", "categories",
    "exchange_rates", "investments", "people",
    "recurring", "transactions", "virtual_accounts",
]
forbidden_modules = ["dashboard", "reports", "push", "jobs", "settings_app"]

[[tool.importlinter.contracts]]
name = "Aggregators must not import from each other"
type = "independence"
modules = ["dashboard", "reports", "push", "jobs", "settings_app"]
